// frontend/src/app/pages/model/DashboardFrag.ts
export class DashboardFrag {
  id: number;
  year: number;
  week: number;
  atelier: string;
  hRecusTrf: number;
  hObjectif: number;
  hRealise: number;
  nBCarteEngage: number;
  nBCarteRetard: number;
  otd: number;
  ecart: number;
  // -------------------------
  mqtTrf: number;
  mqtTa: number;
  qualite: number;
  pcd: number;
  test: number;
  aqf: number;
  bloqueDossier: number;
  maintenance: number;
  processOutillage: number;
  retardProduction: number;
  ecartStock: number;
  attenteInfoClient: number;
  repporterSuiteDeClient: number;
  // --------------------------
  encours: number;
  netFabricable: number;
  enTransit: number;
  monquantTa: number;
  clientMqtFc: number;
  qualiteEncours: number;
  aqfC: number;
  processDossierTra: number;
  testEncours: number;
  machine: number;
  finisAttRiClient: number;
  finisReportDelai: number;
  yM: number;
  // -------------------------------
  mqecartStock: number;
  mqRuptEtConsom: number;
  /// --------------------------------------
  hPrevus: number;
  dontRetard: number;
  fabricable: number;
  volumeFabricableCms: number;
  fabricableEncours: number;
  previsions: number;
  // ----------------------------
  besoinEtp: number;
  etpDispo: number;
  effectifPlus: number;
  // -------------------
  heuresFormation: number;
  heuresAmelioration: number;
  constructor() {}
}